Initial commit: add index, style, and js file.
Finish html layout for the whole project.
All buttons have functionality, i.e.
"+add new book" buttons displays a modal
Modal contains a form where for users to input book's name, author and the number of pages.
It also contains a checkbox for read/not read status.

Upon form submit, dialog closes and all the input data by the user is displayed in a book card on screen.
The read/not read status is not yet shown in book card but this will be added soon.