Fix remove button.
Issue: Actual Book object in myLibrary array was not associated with with DOM Element.
As a result, remove button would only remove the book card from DOM but not from myLibrary array which contains all Book object.

Solution: Upon book card making, an id attribute is added to each book card which corresponds to the index of book object in myLibrary array.